Uploaded the initial version
FaceRecognitise.py uses the FaceRecogniser package to detect a human face and tag the possible matches from the training data. The training pictures for the models with the labels as file names should be provided as the input parameter for the program

This program uses openCV for image operations ( read) and also for capturing the frames from the default webcam of your computer.